added disappeared scooters to trip insensity and filtered out same cluster moved scooters + small refactoring
